refactor into separate files; fix deprecated v8 API use; configure on OSX
