+++ binding.gyp
@@ -2,12 +2,24 @@
   "targets": [
     {
       "target_name": "gstreamer-superficial",
-      "sources": [ "gstreamer.cpp" ],
+      "sources": [ "gstreamer.cpp", "GLibHelpers.cpp", "GObjectWrap.cpp", "Pipeline.cpp" ],
 	  "include_dirs": [
 		"<!(node -e \"require('nan')\")"
 	  ],
 	  "conditions" : [
 		["OS=='linux'", {
+			"include_dirs": [
+				'<!@(pkg-config gstreamer-1.0 --cflags-only-I | sed s/-I//g)',
+				'<!@(pkg-config gstreamer-app-1.0 --cflags-only-I | sed s/-I//g)',
+				'<!@(pkg-config gstreamer-app-1.0 --cflags-only-I | sed s/-I//g)'
+			],
+			"libraries": [
+				'<!@(pkg-config gstreamer-1.0 --libs)',
+				'<!@(pkg-config gstreamer-app-1.0 --libs)',
+				'<!@(pkg-config gstreamer-video-1.0 --libs)'
+			]
+		}],
+		["OS=='mac'", {
 			"include_dirs": [
 				'<!@(pkg-config gstreamer-1.0 --cflags-only-I | sed s/-I//g)',
 				'<!@(pkg-config gstreamer-app-1.0 --cflags-only-I | sed s/-I//g)',
